I2C: Get rid of some build warnings when buidling on 32bit.

+++ src/rhd_i2c.c
@@ -932,8 +932,8 @@ rhdRV620I2CSetupStatus(I2CBusPtr I2CPtr, int line, int prescale)
 
 	reg_7d9c = regList7d9c[line];
 	RHDDebug(I2CPtr->scrnIndex,
-		 "DDC Line[%i] = 0x%8.8x SDL: 0x%2.2x SDA: 0x%2.2x\n",line, reg_7d9c,
-		 reg_7d9c & 0x7F, (reg_7d9c >> 8) & 0x7F) ;
+		 "DDC Line[%i] = 0x%8.8x SDL: 0x%2.2x SDA: 0x%2.2x\n",line, (unsigned int)(reg_7d9c),
+		 (unsigned int)(reg_7d9c & 0x7F), (unsigned int)((reg_7d9c >> 8) & 0x7F)) ;
 	scl_reg = gpioClkReg[line];
     }
